+++ tests/test_get_element.py
@@ -57,6 +57,10 @@ class TestGetElement:
         element = browser.get_element(type="button")
         assert element.selenium_webelement.text == "Add paragraph"
 
+    def test_by_value(self, browser):
+        element = browser.get_element(value="Wallaby")
+        assert element.selenium_webelement.text == "Wallaby"
+
     def test_by_xpath(self, browser):
         element = browser.get_element(xpath="//*[@id='para-2']")
         assert element.selenium_webelement.text == "Paragraph 2"
@@ -110,7 +114,7 @@ class TestGetElementErrors:
         expected = (
             "One parameter from this list is required: class_name, css, id, "
             "link_text, name, partial_link_text, partial_text, tag, text, "
-            "type, xpath"
+            "type, value, xpath"
         )
         assert str(error.value) == expected
 
@@ -145,7 +149,7 @@ class TestGetElementErrors:
         expected = (
             "Only one parameter from this list is allowed: class_name, css, "
             "id, link_text, name, partial_link_text, partial_text, tag, text, "
-            "type, xpath"
+            "type, value, xpath"
         )
         assert str(error.value) == expected