+++ CHANGELOG.rst
@@ -3,6 +3,14 @@ Changelog
 =========
 
 
+Unreleased
+----------
+
+Added
+~~~~~
+- Added "value" kwarg to get_element and get_elements.
+
+
 v0.1.1
 ------
 *11 November 2020*